lab1/requirements.txt
aiogram==3.3.0
python-dotenv==1.0.1
opencv-python-headless==4.10.0.84
pytesseract==0.3.13
